Assignment/requirements.txt
pandas==2.2.3
psycopg2-binary==2.9.10 (dt dec pq3 ext lo64)
openpyxl==3.1.5